feat(native/bridge): refactor hooks into per-file directory structure with web fallbacks
- Extract all 9 hooks from monolithic hooks.js into src/native/bridge/hooks/ (one file per domain)
- hooks.js becomes a 2-line re-export shim for backwards compatibility
- Add web fallbacks for useVideoStream, useCamera, useCameraPermission, useFilePicker,
  useHapticFeedback, useDataProtection, useNetworkStatus, useSafeArea, useDeviceInfo
- Fix useBaseHook isNative detection to use nativeBridge.isAvailable() instead of raw window check
- Fix Rules of Hooks violations in useDataProtection, useFilePicker, useCamera
- Add isNative/isWeb to useSafeArea, useNetworkStatus, useDeviceInfo return shapes
- Fix useNotification interface alignment (add schedule alias, requestPermission, isNative/isWeb)
- Add webFallback prop dynamic sync to all hooks via useEffect
- Update KB with 4 new entries: notification config, sound asset, googleSignIn config, useBaseHook fix
